% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{city2}
\alias{city2}
\title{City distances 2}
\format{A data frame with 78 rows and 3 variables}
\usage{
data(city2)
}
\description{
City distances 2
}
\details{
Distance in miles between 12 cities in the USA. The dataset is used to illustrate multi-dimensional scaling (MDS). Description provided in attr(city2,"description")
}
\keyword{datasets}
