% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{movie}
\alias{movie}
\title{Conjoint data for Movie theaters}
\format{A data frame with 18 rows and 6 variables}
\usage{
data(movie)
}
\description{
Conjoint data for Movie theaters
}
\details{
Rankings reflect the evaluation of 18 alternative movie theaters by one respondent. Description provided in attr(movie,"description")
}
\keyword{datasets}

