% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R, R/pca.R
\name{pca}
\alias{pca}
\title{Principal Component Anlaysis}
\usage{
pca(dataset, vars, scal, cent, pc, data_filter = "", envir = parent.frame())

pca(dataset, vars, scal, cent, pc, data_filter = "", envir = parent.frame())
}
\arguments{
\item{dataset}{Dataset}

\item{vars}{Vector of variables to include in the analysis}

\item{scal}{Boolean value if the dataset is to be scaled}

\item{cent}{Boolean value if the dataset is to be centered}

\item{pc}{Maximum principal components (default is equal to the number of variables)}

\item{data_filter}{Expression entered in, e.g., Data > View to filter the dataset in Radiant. The expression should be a string (e.g., "price > 10000")}

\item{envir}{Environment to extract data from}

\item{pca_scale}{Boolean value if the dataset is to be scaled}

\item{pca_center}{Boolean value if the dataset is to be centered}

\item{pca_pc}{Maximum principal components (default is equal to the number of variables)}

\item{plotpc}{Boolean value if the PCA is to be visualized}
}
\value{
A list of all variables used in pca as an object of class pca

A list of all variables used in pca as an object of class pca
}
\description{
Principal Component Anlaysis

Principal Component Analysis
}
\details{
See \url{https://radiant-rstats.github.io/docs/multivariate/hclus.html} for an example in Radiant

Principal Component Analysis ....
}
\examples{
pca(shopping, vars = "v1:v6") \%>\% str()

}
\seealso{
\code{\link{summary.hclus}} to summarize results

\code{\link{plot.hclus}} to plot results
}
