% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_factor.R
\name{plot.full_factor}
\alias{plot.full_factor}
\title{Plot method for the full_factor function}
\usage{
\method{plot}{full_factor}(x, shiny = FALSE, custom = FALSE, ...)
}
\arguments{
\item{x}{Return value from \code{\link{full_factor}}}

\item{shiny}{Did the function call originate inside a shiny app}

\item{custom}{Logical (TRUE, FALSE) to indicate if ggplot object (or list of ggplot objects) should be returned. This opion can be used to customize plots (e.g., add a title, change x and y labels, etc.). See examples and \url{http://docs.ggplot2.org/} for options.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the full_factor function
}
\details{
See \url{https://radiant-rstats.github.io/docs/multivariate/full_factor.html} for an example in Radiant
}
\examples{
result <- full_factor("diamonds",c("price","carat","table"))
plot(result)
result <- full_factor("computer","high_end:business")
summary(result)

}
\seealso{
\code{\link{full_factor}} to calculate results

\code{\link{plot.full_factor}} to plot results
}
