% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kclus.R
\name{plot.kclus}
\alias{plot.kclus}
\title{Plot method for kclus}
\usage{
\method{plot}{kclus}(x, plots = "density", shiny = FALSE, custom = FALSE,
  ...)
}
\arguments{
\item{x}{Return value from \code{\link{kclus}}}

\item{plots}{One of "density", "bar", or "scatter")}

\item{shiny}{Did the function call originate inside a shiny app}

\item{custom}{Logical (TRUE, FALSE) to indicate if ggplot object (or list of ggplot objects) should be returned. This opion can be used to customize plots (e.g., add a title, change x and y labels, etc.). See examples and \url{http://docs.ggplot2.org/} for options.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for kclus
}
\details{
See \url{https://radiant-rstats.github.io/docs/multivariate/kclus.html} for an example in Radiant
}
\examples{
result <- kclus("shopping", vars = c("v1:v6"))
plot(result)
shopping \%>\% kclus(, vars = c("v1:v6")) \%>\% plot

}
\seealso{
\code{\link{kclus}} to generate results

\code{\link{summary.kclus}} to summarize results

\code{\link{store.kclus}} to add cluster membership to the selected dataset
}
