% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot.pca}
\alias{plot.pca}
\title{Plot method for the pca function}
\usage{
\method{plot}{pca}(x, plots = c("scree", "biplot"), shiny = FALSE, custom = FALSE, ...)
}
\arguments{
\item{x}{Return value from \code{\link{pca}}}

\item{plots}{Plots to return.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the pca function
}
\details{
Plotting a PCA ...
}
\seealso{
\code{\link{hclus}} to generate results

\code{\link{summary.hclus}} to summarize results
}
