% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjoint.R
\name{print.conjoint.predict}
\alias{print.conjoint.predict}
\title{Print method for predict.conjoint}
\usage{
\method{print}{conjoint.predict}(x, ..., n = 20)
}
\arguments{
\item{x}{Return value from prediction method}

\item{...}{further arguments passed to or from other methods}

\item{n}{Number of lines of prediction results to print. Use -1 to print all lines}
}
\description{
Print method for predict.conjoint
}
