% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{package}
\name{radiant.multivariate-deprecated}
\alias{radiant.multivariate-deprecated}
\alias{save_factors}
\alias{save_membership}
\alias{kmeans_clus}
\alias{hier_clus}
\alias{radiant.multivariate-deprecated-package}
\title{Deprecated function(s) in the radiant.multivariate package}
\usage{
save_factors(...)
}
\arguments{
\item{...}{Parameters to be passed to the updated functions}
}
\description{
These functions are provided for compatibility with previous versions of
radiant. They will eventually be  removed.
}
\section{Details}{

\tabular{rl}{
  \code{save_factors} is now a synonym for \code{\link{store.full_factor}}\cr
  \code{save_membership} is now a synonym for \code{\link{store.kclus}}\cr
  \code{kmeans_clus} is now a synonym for \code{\link{kclus}}\cr
  \code{hier_clus} is now a synonym for \code{\link{hclus}}\cr
}
}

