% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/radiant.R
\docType{package}
\name{radiant.multivariate}
\alias{radiant.multivariate}
\alias{radiant.multivariate-package}
\alias{radiant.multivariate}
\title{radiant.multivariate}
\usage{
radiant.multivariate()
}
\description{
radiant.multivariate

Launch Radiant in the default browser
}
\details{
See \url{https://radiant-rstats.github.io/docs} for documentation and tutorials
}
