% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{retailers}
\alias{retailers}
\title{Perceptions of retailers}
\format{
A data frame with 6 rows and 10 variables
}
\usage{
data(retailers)
}
\description{
Perceptions of retailers
}
\details{
Consumer evaluations for a set of retailers in the Chicago area on 7 attributes. The dataset is used to illustrate perceptual maps. Description provided in attr(retailers, "description")
}
\keyword{datasets}
