% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjoint.R
\name{store.conjoint.predict}
\alias{store.conjoint.predict}
\title{Store predicted values generated in predict.conjoint}
\usage{
\method{store}{conjoint.predict}(object, ..., data = attr(object,
  "pred_data"), name = "prediction")
}
\arguments{
\item{object}{Return value from model predict function}

\item{...}{Additional arguments}

\item{data}{Data or dataset name (e.g., data = mtcars or data = "mtcars")}

\item{name}{Variable name(s) assigned to predicted values}
}
\description{
Store predicted values generated in predict.conjoint
}
\details{
See \url{https://radiant-rstats.github.io/docs/multivariate/conjoint.html} for an example in Radiant
}
