% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjoint.R
\name{store.conjoint}
\alias{store.conjoint}
\title{Store method for the Multivariate > Conjoint tab}
\usage{
\method{store}{conjoint}(dataset, object, name, ...)
}
\arguments{
\item{dataset}{Dataset}

\item{object}{Return value from conjoint}

\item{name}{Variable name(s) assigned to predicted values}

\item{...}{further arguments passed to or from other methods}
}
\description{
Store method for the Multivariate > Conjoint tab
}
\details{
Store data frame with PWs or IWs in Radiant r_data list if available
}
