% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_factor.R
\name{store.full_factor}
\alias{store.full_factor}
\title{Store factor scores to active dataset}
\usage{
\method{store}{full_factor}(dataset, object, name = "", ...)
}
\arguments{
\item{dataset}{Dataset to append to factor scores to}

\item{object}{Return value from \code{\link{full_factor}}}

\item{name}{Name of factor score variables}

\item{...}{Additional arguments}
}
\description{
Store factor scores to active dataset
}
\details{
See \url{https://radiant-rstats.github.io/docs/multivariate/full_factor.html} for an example in Radiant
}
\examples{
full_factor(shopping, "v1:v6", nr_fact = 3) \%>\%
  store(shopping, .) \%>\%
  head()

}
\seealso{
\code{\link{full_factor}} to generate results

\code{\link{summary.full_factor}} to summarize results

\code{\link{plot.full_factor}} to plot results
}
