% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kclus.R
\name{store.kclus}
\alias{store.kclus}
\title{Add a cluster membership variable to the active dataset}
\usage{
\method{store}{kclus}(dataset, object, name = "", ...)
}
\arguments{
\item{dataset}{Dataset to append to cluster membership variable to}

\item{object}{Return value from \code{\link{kclus}}}

\item{name}{Name of cluster membership variable}

\item{...}{Additional arguments}
}
\description{
Add a cluster membership variable to the active dataset
}
\details{
See \url{https://radiant-rstats.github.io/docs/multivariate/kclus.html} for an example in Radiant
}
\examples{
kclus(shopping, vars = "v1:v6", nr_clus = 3) \%>\%
  store(shopping, .) \%>\%
  head()
}
\seealso{
\code{\link{kclus}} to generate results

\code{\link{summary.kclus}} to summarize results

\code{\link{plot.kclus}} to plot results
}
