% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_clus.R
\name{store.kmeans_clus}
\alias{store.kmeans_clus}
\title{Add a cluster membership variable to the active dataset}
\usage{
\method{store}{kmeans_clus}(object, ..., name = "")
}
\arguments{
\item{object}{Return value from \code{\link{kmeans_clus}}}

\item{...}{Additional arguments}

\item{name}{Name of cluster membership variable}
}
\description{
Add a cluster membership variable to the active dataset
}
\details{
See \url{https://radiant-rstats.github.io/docs/multivariate/kmeans_clus.html} for an example in Radiant
}
\examples{
kmeans_clus(shopping, vars = c("v1:v6")) \%>\% store \%>\% head

}
\seealso{
\code{\link{kmeans_clus}} to generate results

\code{\link{summary.kmeans_clus}} to summarize results

\code{\link{plot.kmeans_clus}} to plot results
}

