% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclus.R
\name{summary.hclus}
\alias{summary.hclus}
\title{Summary method for the hclus function}
\usage{
\method{summary}{hclus}(object, ...)
}
\arguments{
\item{object}{Return value from \code{\link{hclus}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the hclus function
}
\details{
See \url{https://radiant-rstats.github.io/docs/multivariate/hclus.html} for an example in Radiant
}
\examples{
result <- hclus("shopping", vars = c("v1:v6"))
summary(result)

}
\seealso{
\code{\link{hclus}} to generate results

\code{\link{plot.hclus}} to plot results
}
