% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kclus.R
\name{summary.kclus}
\alias{summary.kclus}
\title{Summary method for kclus}
\usage{
\method{summary}{kclus}(object, dec = 2, ...)
}
\arguments{
\item{object}{Return value from \code{\link{kclus}}}

\item{dec}{Number of decimals to show}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for kclus
}
\details{
See \url{https://radiant-rstats.github.io/docs/multivariate/kclus.html} for an example in Radiant
}
\examples{
result <- kclus(shopping, vars = "v1:v6", nr_clus = 3)
summary(result)

}
\seealso{
\code{\link{kclus}} to generate results

\code{\link{plot.kclus}} to plot results

\code{\link{store.kclus}} to add cluster membership to the selected dataset
}
