% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_clus.R
\name{summary.kmeans_clus}
\alias{summary.kmeans_clus}
\title{Summary method for kmeans_clus}
\usage{
\method{summary}{kmeans_clus}(object, dec = 2, ...)
}
\arguments{
\item{object}{Return value from \code{\link{kmeans_clus}}}

\item{dec}{Number of decimals to show}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for kmeans_clus
}
\details{
See \url{https://radiant-rstats.github.io/docs/multivariate/kmeans_clus.html} for an example in Radiant
}
\examples{
result <- kmeans_clus("shopping", vars = c("v1:v6"))
summary(result)
shopping \%>\% kmeans_clus(vars = c("v1:v6"), nr_clus = 3) \%>\% summary

}
\seealso{
\code{\link{kmeans_clus}} to generate results

\code{\link{plot.kmeans_clus}} to plot results

\code{\link{store.kmeans_clus}} to add cluster membership to the selected dataset
}

