% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{summary.mds}
\alias{summary.mds}
\title{Summary method for the mds function}
\usage{
\method{summary}{mds}(object, dec = 2, ...)
}
\arguments{
\item{object}{Return value from \code{\link{mds}}}

\item{dec}{Rounding to use for output (default = 2). +1 used for stress measure}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the mds function
}
\details{
See \url{https://radiant-rstats.github.io/docs/multivariate/mds.html} for an example in Radiant
}
\examples{
result <- mds(city, "from", "to", "distance")
summary(result, dec = 1)

}
\seealso{
\code{\link{mds}} to calculate results

\code{\link{plot.mds}} to plot results
}
