% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R, R/pca.R
\name{summary.pca}
\alias{summary.pca}
\title{Summary method for the pca function}
\usage{
\method{summary}{pca}(object, ...)

\method{summary}{pca}(object, ...)
}
\arguments{
\item{object}{Return value from \code{\link{pca}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the pca function

Summary method for the pca function
}
\details{
See \url{https://radiant-rstats.github.io/docs/multivariate/hclus.html} for an example in Radiant

See \url{https://radiant-rstats.github.io/docs/multivariate/hclus.html} for an example in Radiant
}
\examples{
result <- summary.pca(shopping, vars = c("v1:v6"))
summary(result)

}
\seealso{
\code{\link{hclus}} to generate results

\code{\link{plot.hclus}} to plot results
}
