% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_factor.R
\name{summary.pre_factor}
\alias{summary.pre_factor}
\title{Summary method for the pre_factor function}
\usage{
\method{summary}{pre_factor}(object, dec = 2, ...)
}
\arguments{
\item{object}{Return value from \code{\link{pre_factor}}}

\item{dec}{Rounding to use for output}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the pre_factor function
}
\details{
See \url{https://radiant-rstats.github.io/docs/multivariate/pre_factor.html} for an example in Radiant
}
\examples{
result <- pre_factor(shopping, "v1:v6")
summary(result)
pre_factor(computer, "high_end:business") \%>\% summary()
}
\seealso{
\code{\link{pre_factor}} to calculate results

\code{\link{plot.pre_factor}} to plot results
}
