% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{tpbrands}
\alias{tpbrands}
\title{Toothpaste brands}
\format{A data frame with 45 rows and 4 variables}
\usage{
data(tpbrands)
}
\description{
Toothpaste brands
}
\details{
Perceived (dis)similarity of a set of toothpaste brands. The dataset is used to illustrate multi-dimensional scaling (MDS). Description provided in attr(tpbrands,"description")
}
\keyword{datasets}
