% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R, R/pca.R
\name{pca}
\alias{pca}
\title{Principal Component Anlaysis}
\usage{
pca(dataset, pca_scale, pca_center, pca_pc)

pca(dataset, pca_scale, pca_center, pca_pc)
}
\arguments{
\item{dataset}{Dataset}

\item{pca_scale}{Boolean value if the dataset is to be scaled}

\item{pca_center}{Boolean value if the dataset is to be centered}

\item{pca_pc}{Maximum principal components (default is equal to the number of variables)}

\item{pca_plot}{Boolean value if the PCA is to be visualized}
}
\value{
A list of all variables used in pca as an object of class pca

A list of all variables used in pca as an object of class pca
}
\description{
Principal Component Anlaysis

Principal Component Analysis
}
\details{
See \url{https://radiant-rstats.github.io/docs/multivariate/hclus.html} for an example in Radiant

Principal Component Analysis ....
}
\examples{
pca(shopping, vars = "v1:v6") \%>\% str()

}
\seealso{
\code{\link{summary.hclus}} to summarize results

\code{\link{plot.hclus}} to plot results
}
