% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{package}
\name{radiant.multivariate-deprecated}
\alias{radiant.multivariate-deprecated}
\alias{pmap}
\alias{radiant.multivariate-deprecated-package}
\title{Deprecated function(s) in the radiant.multivariate package}
\usage{
pmap(...)
}
\arguments{
\item{...}{Parameters to be passed to the updated functions}
}
\description{
These functions are provided for compatibility with previous versions of
radiant. They will eventually be  removed.
}
\section{Details}{

\itemize{
  \item Replace \code{pmap} by \code{\link{prmap}}
}
}

