% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/radiant.R
\docType{package}
\name{radiant.multivariate}
\alias{radiant.multivariate}
\alias{radiant.multivariate-package}
\title{radiant.multivariate}
\usage{
radiant.multivariate(state, ...)
}
\arguments{
\item{state}{Path to state file to load}

\item{...}{additional arguments to pass to shiny::runApp (e.g, port = 8080)}
}
\description{
radiant.multivariate

Launch radiant.multivariate in the default browser
}
\details{
See \url{https://radiant-rstats.github.io/docs} for documentation and tutorials
}
\examples{
\dontrun{
radiant.multivariate()
}
}
