% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjoint.R
\name{store.conjoint.predict}
\alias{store.conjoint.predict}
\title{Store predicted values generated in predict.conjoint}
\usage{
\method{store}{conjoint.predict}(dataset, object, name = "prediction", ...)
}
\arguments{
\item{dataset}{Dataset to add predictions to}

\item{object}{Return value from model predict function}

\item{name}{Variable name(s) assigned to predicted values}

\item{...}{Additional arguments}
}
\description{
Store predicted values generated in predict.conjoint
}
\details{
See \url{https://radiant-rstats.github.io/docs/multivariate/conjoint.html} for an example in Radiant
}
\examples{
conjoint(mp3, rvar = "Rating", evar = "Memory:Shape") \%>\%
  predict(mp3) \%>\%
  store(mp3, ., name = "pred_pref")

}
