% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjoint.R
\name{store.conjoint}
\alias{store.conjoint}
\title{Store method for the Multivariate > Conjoint tab}
\usage{
\method{store}{conjoint}(object, name = "PWs", type = "PW",
  envir = parent.frame(), ...)
}
\arguments{
\item{object}{Return value from conjoint}

\item{name}{Name of the dataset to store}

\item{type}{Type of output to store}

\item{envir}{Environment to assign 'new' dataset (optional). Used when an r_data list is not available}

\item{...}{further arguments passed to or from other methods}
}
\description{
Store method for the Multivariate > Conjoint tab
}
\details{
Store data frame with PWs or IWs in Radiant r_data list if available
}
